Clean up native usage follow-up
