+++ src/benchflow/providers/litellm_runtime.py
@@ -25,10 +25,7 @@ import httpx
 import yaml
 
 from benchflow.agents.codex_config import apply_codex_provider_config
-from benchflow.agents.env import (
-    agent_supports_native_acp_usage,
-    uses_native_subscription_auth,
-)
+from benchflow.agents.env import uses_native_subscription_auth
 from benchflow.agents.registry import AGENTS
 from benchflow.providers.litellm_config import (
     LITELLM_MASTER_KEY_ENV,
@@ -42,10 +39,9 @@ from benchflow.providers.litellm_logging import (
     callback_module_source,
     extract_usage_from_trajectory,
     trajectory_from_litellm_callback_log,
-    usage_unavailable,
 )
 from benchflow.trajectories.types import Trajectory
-from benchflow.usage_tracking import UsageTrackingConfig
+from benchflow.usage_tracking import UsageTrackingConfig, usage_unavailable
 
 logger = logging.getLogger(__name__)
 
@@ -966,19 +962,10 @@ async def ensure_litellm_runtime(
         )
 
     if uses_native_subscription_auth(agent, model, agent_env):
-        if usage_cfg.mode == "required" and not agent_supports_native_acp_usage(agent):
-            raise RuntimeError(
-                "Token usage tracking is required, but agent "
-                f"{agent!r} cannot report native ACP usage."
-            )
         return await _skip_litellm_runtime(
             agent_env,
             runtime,
-            reason=(
-                "native subscription auth will use agent ACP usage telemetry"
-                if agent_supports_native_acp_usage(agent)
-                else "native subscription auth bypasses LiteLLM"
-            ),
+            reason="native subscription auth will use agent ACP usage telemetry",
         )
 
     if not needs_litellm_runtime(agent, model):
